requirements.txt
aiohttp>=3.9.3
PyYAML>=6.0.2